topcat: fix ColumnInfoWindow metadata display regression bug
Per-column TopcatModel metadata is diplayed in a JTable in
the ColumnInfoWindow.  When a new non-empty metadata info appears
(e.g. a new TopcatModel table column is added with an Expression,
when none of the columns previously had an Expression),
then the relevant JTable column is supposed to be revealed
(unhidden) automatically.  This was only happening when the
*second* column with the new metadata item was added.

The problem is to do with the sequence in which model update events
are fired between the various different models - I haven't nailed
down the exact details.  To fix it, I've used SwingUtilities.invokeLater
in one of the event handlers to queue a subsequent event.
This is not particularly respectable, and there is no doubt a cleaner way,
but it does the job, and should (unlike a cleaner way) be robust
against this sort of thing happening again.

This was a regression bug since topcat v4.0b (actually introduced
at a98933ea9015).  Before that it wasn't showing up because adding
a new column first added the column and then moved it, which meant
that the model had fireTableDataChanged() called on it twice in
quick succession, so though it missed it the first time it picked
it up on the second.  So it was working earlier by accident,
rather than the change doing something wrong.

Thanks to Doug Burke for reporting this.