topcat: fix shader control to work per-zone
The shader control is no longer passed from StackPlotWindow into
PlotPanel, the StackPlotWindow acquires the relevant per-zone
information from the ShaderControl and passes that on.
This is cleaner, and it allows the shading to be configured
differently for each zone (though currently it's all controlled
by the same ShaderControl GUI).