topcat: AxisController no longer contains StackModel
The GUI now passes relevant layer controls to AxisController instances
when they may have changed, rather than the AxisController keeping
a copy of the layer stack to update on its own.  This makes control
flow more transparent, and also makes it possible to configure
AxisController instances on a per-zone basis.