vo: fix sky join example syntax in ADQL Hints panel
The example text for the sky join (CONTAINS(POINT,CIRCLE)) was wrong -
it had a closing parenthesis and comma in the wrong place.  Ouch!
That was a pretty nasty mistake to make.  Interesting that nobody
had reported it.  Obviously the Hints panel is not such a super-useful
addition to the TAP window as I thought.