topcat: AxisControllers are now managed per-zone
Multi-zone plots have one AxisController per zone, managed by a
MultiController object.  This currently inserts one Axes fixed control
into the control stack for each currently visible zone.
That's not a particularly nice GUI, but it works, and the framework
should form a stable API on which to build a nicer GUI.