topcat: better provision of ZoneIds
ZoneIds are now obtained from a ZoneFactory; you (currently)
get a ZoneFactory from a PlotTypeGui (maybe a PlotType in future
if I find I need them in stilts too).  This means it's possible
to control what ZoneIds are in play, allowing them to be ordered
without error, and also gives more explicit ways to indicate
things like whether the context is single- or multi-zone.