ttools: time stack now goes upwards not downwards
Add a constructor parameter in StackGanger to determine whether
later zones in the sequence appear above or below earlier ones.
From a java point of view, up makes sense, but from a science
point of view, down makes sense.  Which sounds like it should
go down, but then there are GUI components like the JComboBox
where the sequence is downwards.  Anyway, punt the decision
in StackGanger, but in TimeStackGanger choose down (for now).