ttools: spectrogram plotter now uses shared (aux) colormap
The colour map used by the spectrogram plotter is now the shared
map for the plot (in fact the zone), rather than one private to
its layer.  That means that the plotting infrastructure will
display the colour ramp and allow it to be manipulated with a
shading control or whatever else is usually used to tweak the map.

So we now have a shading control in topcat to let you manipulate
spectrogram colour, and a shade axis (visible colour ramp)
to see quantitatively what colour mean.  This is probably a good
thing, but it may be less intuitive to find it per-zone rather
than per-layer when configuring it.  So possibly back out of this
in the future.