ttools: get FunctionPlotter (barely) working in TimePlot
FunctionPlotter was previously causing a ClassCastException to be
thrown when used with the Time plot - it was trying to cast the
Surface to a PlaneSurface.  This isn't necessary, so I removed the
offending lines and now you can include function layers in the time plot.

They are not very user-friendly though, since the X coordinate is
the internal time plot time coordinate - unix seconds.  If you don't
know or can't work that out, you can't do sensible plots against
time, and even if you do/can, it's fiddly.  Since time plots are still
considered experimental, leave this in place, but for proper
time function plotting it should get improved in the future.