ttools: fix FillPlotter display between existing points
Add more information to the fill layer data structure (FillPlan)
recording the closest points to the pixel grid that fall outside
that grid.  This is necessary to work out where the fill level
comes up to if the plotted region does not contain any data points
at all.

This information could be used to provide slightly better behaviour
near the edges of the plot than is currently iplemented, but it's
not too noticeable.