table: fix numeric field truncation bug in LaTeX table output
LatexTableWriter was truncating numeric fields to 16 characters -
although there are approximately 16 sig figs in double precision
floating point numbers, that doesn't account for decimal points,
sign characters and exponents.  The result was an ellipsis at the
end which could hide the exponent - oops.  Fixed.
The LaTeX tabular environment output still is not very sophisticated,
but at least it's more correct.

Thanks to Natasha Hurley-Walker for reporting this.