ttools: add a couple of assertions in axis labelling code
BasicTicker assertions.  Tick generation is tricky, it needs all
the help it can get.  However, these assertions are a bit tricky,
since there are precision issues to consider, especially for use
with time axes, where the absolute index values may be quite large
(near 15 decimal sig figs, the limit of doubles).  So guard them
with checks on the absolute value of the index.