ttools: rewrite ISO-8601 tickmark generation
There was a bug in the TimeFormat implementation that meant ISO-8601
tickmarks were sometimes drawn in badly incorrect places.
The whole way the DateRule was implemented was wrongly thought out,
so it needed a rewrite.  I think it's OK now, but this stuff is
very tricky, so keep an eye on it.