ttools: changed GraphicExporter.GIF encoder implementation
The GraphicExporter.GIF implementation now uses the JRE's ImageIO
rather than the external Acme.JPM.Encoders.GifEncoder.
It can do this because as of java 1.6, ImageIO contains a GIF writer
(previously it didn't because of patent issues).
From a quick look, the results are pretty similar.
It's possible there are discrepancies, but GIF is a bit of a
minority interest these days, so I don't expect it to cause
trouble even if that's the case.

The external Acme library can be removed as a dependency of ttools,
which saves bytes and complication.

The Acme GifEncoder is still a dependency for topcat.  That should
be removed at some point, since topcat should really be using PNGs
where it's currently using GIFs.