topcat: remove alt.java hack for documentation builds
The topcat and ttools packages build user documents from large XML
source files.  This was hitting xerces bugs at build time when
building with Sun/Oracle J2SE 1.5, so there was a hack in place
in the ant build files that let you set the alt.java system property
to use a different JDK when building the documentation from building
the classes.

Since the build environment is now guaranteed to be java 1.6+,
this hack is no longer required.  Excise it.