ttools: taplint now checks ObsCore 1.1
ObsTapStage now looks for declaration of the ObsCore 1.1 data model
and makes specific checks for it.

This matches the specifications in PR-ObsCore-v1.1-20160330.

That version of the standard claims that ObsCore 1.1 is a
strict superset of ObsCore 1.0, which it isn't because of some
changes to the Utypes and UCDs between the two.  If that anomaly
is resolved in the document, or other changes are made before REC,
it may be necessary to revise this code to match.