ttools: rename 2-arg Arithmetic.max,min -> maxNaN,minNaN
The 2-argument max/min functions provided by the Arithmetic class
for the expression language are renamed to maxNaN and minNaN.
This is because vararg functions max,min are provided in the Lists
class, with slightly different behaviour (when one of the arguments
is a NaN).

This does mean a change in behaviour at the user level, but probably
nobody actually wants this NaN-preserving behaviour anyway.