ttools: HealpixPlotter aux ranging now sensitive to surface
The Aux axis range generated by HealpixPlotter is now determined
by what pixels are visible in the plot (as for other plot types)
rather than globally including invisible pixels.

This is not very efficient though - the aux ranging is now likely
to be more expensive (it requires an iteration over the data)
than the actual painting.