ttools: HealpixPlotter paints not resamples big tiles
Implement a choice of strategies for drawing healpix tiles.
If the tiles are large (compared to a screen pixel), they are
painted using AWT graphics prmimitives, and if they are small
they are resampled.  This makes plotting much faster (pretty
fast in fact) if the tiles are large, especially if the plot
is large in pixel count.

This works ... but not very well.  Quite often some tiles are not
painted.  This is evidently down to the code in SkySurfaceTiler
that identifies which healpix tiles are visible on a given
plotting surface.  I suspect a problem with PixTools.query_polygon
but possibly the bug is mine.

PixTools is in any case regularly complaining:

   The polygon has more than one concave vertex
   The result is unpredictable