fits: fix read failure for TDIM with zero-length columns
In the case that a FITS file has a zero-length column
(e.g. TFORMn = 0E) and a corresponding non-blank TDIMn header
(e.g. TDIMn = (0,10)), reading the file was generating an
IllegalArgumentException in DefaultValueInfo.setShape(), since
that requires strictly positive values for all shape dimensions.
This was encountered in some Kepler (K2-C3) files, e.g.
http://archive.stsci.edu/pub/k2/target_pixel_files/c3/205900000/20000/
                         ktwo205920873-c03_lpd-targ.fits.gz

FITS does not disallow zero-length TDIMn elements (except for
variable-length arrays), but I doubt if it will cause trouble
not setting shape metadata in the StarTable in this special case,
and it might cause trouble to allow zero elements in the StarTable
shape array, so just avoid setting the shape in this case.