ttools: write Examples stage for taplint
New taplint class ExampleStage tests the DALI/TAP examples endpoint
for a TAP service.  This is currently based mostly on the following
drafts:

   WD-DALI-1.1-20160415
   WD-TAP-1.1-20160428

somewhat informed by the older

   TAPNotes-2013-12-13

so changes may be required as those standards become finalised.

The implementation is fairly complete, though it does not currently
test registration of the examples capability in the capabilities file.