topcat: document new time plot changes in SUN/253
Update the section on the Time Plot and Spectrogram controls
in accordance with recent time-related changes to the code.
The Time Plot is still noted as experimental, but there is
now discussion of plot Zones (including a new section)
and the spectrogram now uses the Aux control for colour map
configuration rather than having its own per-layer style options.

This commit does not include updates to the change log.