ttools: modify auto-documentation for plot2task zone suffixes
AbstractPlot2Task and friends now provide the documentation versions
of zone-dependent parameters without the trailing "Z", even for
multi-zone tasks (currently, only plot2time).  This is probably
more comprehensible, and certainly will be for tasks where a
single zone is the norm.  In compensation, the detailed descriptions
of the relevant parameters contain explicit discussion of how
to use the zone suffixes.