ttools: improve licensing documentation
Write a file LICENSE.txt, distributed in stilts.jar and stilts_jars.zip,
which explains the licensing arrangements.  Basically, the STILTS
application is encumbered with the GPL since some of its dependencies
are, but most of the library code, including all of starjava, can
be used under more permissive licenses.

Also drop README.cea from the distributed jar files, since it's
extremely unlikely to be of interest to anybody now.