topcat: reorganise recent version history items etc
Rerrange the items in the SUN/253 version history for the upcoming
release into a more logical sequence.

Plus some other miscellanea:
adjust some comments about the stacked line plot elsewhere
in the document, since the Time plot now does this, and
add a missing icon to the section on the Fill plot2 form.