ttools: make the colourmap boost configurable
The colour maps provided by StyleKeys are doctored so that they
are distinguishable from white over their whole range, so that
when used to colour isolated markers in a scatter plot you can
still see those markers.  This is done by configuring a "boost"
in the clip at the low end of the ramp for ramps that fade to white.

However, that's not such a good idea for cases where the whole
plotting surface is coloured in, e.g. density maps and spectrograms.
Daniel Michalik in particular has complained about it, since he
wants the whole range of certain particular colour ramps when
generating (Gaia) TGAS density maps for publication.

This commit allows the existing Clip option to accept a null value
as an alternative to an explicit (lo,hi) Subrange.  Such a null
value is interpreted by a RampKeySet as a request to use the default
subrange; such defaults can be configured per-colourmap to include
a boost.  If you want the whole range, you can specify clip=0,1
explicitly.  In the GUI this is represented by a toggle button
labelled "Default" alongside the clip subrange double-slider,
which toggles between a null and non-null config value.

I initially considered just adding a separate boolean Boost config
option, which would affect how the existing Clip option was interpreted.
While that would be conceptually cleaner, it would also have
meant introducing (yet) another config option, not much used and
not easily understood, to clutter up the UI for colour map selection.