ttools: SkyDensityPlotter now Reports an export StarTable
SkyDensityPlotter includes in its ReportMap a StarTable containing
the content of the HEALPix map currently plotted.  Creating the
map in this way is not expensive because of the way that plotter
operates.

This could be used by harness applications to give users the
opportunity to export the HEALPix plot they are viewing for later use.

This commit also declares that SkyDensityPlotter hasReports,
which means the tile size will become visible in reports.
Probably that should have been done anyway.