topcat: don't default certain table exports to ascii
The AuxWindow.createSaveTableAction until now has defaulted to
ASCII table output "because it's typically used for small tables".
That default dates from when it was initially used for export of
data from the Statistics window.

Now it's not necessarily small tables or ones that do not benefit
from (e.g. FITS[-plus]) metadata, so remove that default, and
use the usual "(auto)" instead.  The save format is still
sensitive to the user-supplied file extension, so although
this is a change in behaviour, it's not likely to be much noticed.