vo: implement latest VOSI1.1 proposal for tableset retrieval
In updates to the VOSI 1.1 PR agreed in Cape Town and added to the
20160414 PR on 6 July 2016, the /tables endpoint will accept both
detail=min and detail=max, both of them ignorable hints.

Update the VosiTapMetaReader accordingly, and add a new VOSI 1.1
TapMetaPolicy option to allow all the possible hints to be conveyed
by choosing different Metadata Acquisition policies from (e.g.)
the topcat TAP menu.

Also change the implementation of the AUTO policy.  This now uses
TAP_SCHEMA if there are many tables or VOSI1.1 detail=max if there
are few.

This new policy should now work with pretty much all VOSI 1.0 and 1.1
services.  For that reason once the protocol is finalised I can
probably retire the CadcTapMetaReader, TableSetTapMetaReader and
VizierTapMetaReader implementations.  Note however that the the
VOSI 1.1 reader does not work with the (non-standard) TAPVizieR
service, which at time of writing requires schema-qualified names
in the detail path.