topcat: use ControlWindow for report export actions
The ReportPanel uses AuxWindow.create{Save,Import}Action instance
methods for the layer data export actions.  It needs an AuxWindow
instance to call these.  Previously it was looking for one in the
Swing ancestor hierarchy, but if the component was not displayed
because another layer control was selected, that doesn't work,
because it has no window ancestor.  Instead, use the ControlWindow
instance.  The behaviour is slightly different, since the save window
will now be shared among all plot windows, but probably that is
preferable.