ttools: fix documentation bug in TilingParameter
TilingParameter's auto-documentation has reported in all public
releases to date that it takes values of the form
"healpix<scheme><nside>" when in fact it takes values of the form
"healpix<scheme><k>".  The latter is much more sensible,
so change the documentation to match the (historical and current)
behaviour.

Evidently nobody has ever actually used this parameter, or at
least paid attention to its documentation.  That's not surprising,
since it has till now only been used for the rather special-interest
task sqlskymatch.