ttools: add JELUtils.compileQuantity method
The new method JELUtils.compileQuantity and the JELQuantity class
that goes with it, allows you to recover metadata from table columns
if it's there (if your JEL expression essentially denotes an existing
column of a table).  A new JELTable factory method goes with it.

In principle this should allow improved metadata propagation in
stilts commands and elsewhere, however at this commit it's not
actually used anywhere.