fits: fix bug causing 32-bit large FITS file read failures
There was a silly bug in the FITS handling code that used the
getName() method of java.io.File when it should have used getPath().
That meant that in certain cases files were not found if they
were located outside of the current working directory.

This code was only encountered when looking at large (>0.5Gb)
FITS files and running on a 32-bit JVM, which is why it had
gone unnoticed for a long time.

It was a regression bug since STIL v3.0-9/STILTS v3.0-2/TOPCAT v4.2-2.

Thanks to John Pye from Leicester for reporting this.