ttools: use TapEndpointParams in tap{lint,query,skymatch}
Replace use of a URLParameter with TapEndpointParams in the STILTS
tasks taplint, tapquery and tapskymatch.

This means that you can use task parameters to specify custom
locations for the various TAP endpoints, rather than having them
fixed at their default locations hanging off the base TAP service URL.

This is useful for taplint where you may want to mix and match,
especially for a partially deployed service.
For tapquery and tapskymatch, it's currently less useful, since only
the sync and async endpoints are actually used (I think).
So for those latter two commands, for now don't list the extra
parameters in the parameter lists, leaving them largely undocumented.
But the extra parameters are there if required by expert users,
for instance in order to talk to async query endpoints with
non-standard securityMethods.