ttools: AuxReader interface no longer takes TupleSequence
AuxReader.adjustAuxRange now takes a DataSpec,DataStore pair
(i.e. a TupleSequence factory) rather than a TupleSequence.
This gives it more flexibility: it can create as many or
few TupleSequences as it needs to in order to do its job rather
than being stuck with exactly one.

The current motivation is that in some cases it may be able to
get away with no data scans at all, so it doesn't want to be
presented with a tuple sequence it may never use.

At this commit though, no advantage is taken of this new flexibility,
so there's no change.