ttools: make known plot plans available for range calcs
The AuxReader interface's adjustAuxRange method is modified to
accept a knownPlans object.  This may be empty and may be ignored,
but it gives the plotting harness the chance to make planning
information available to AuxReaders when they are trying to
determine range information, which may mean they can avoid work.

At this commit, no useful work is done using this new feature.