ttools: HealpixPlotter uses plans for ranging
The recent plot2 API changes mean that previously calculated
plot plans may be available to the HealpixPlotter when it is trying
to calculate pixel value ranges.  If so, it can use this information
to avoid having to do another scan of the data at range time.
That should make the plot refresh cycle much faster where
applicable (topcat, but not stilts, since stilts currently does
not re-range on replots, so it's not slow anyway).