topcat: adjust which plot surface is used for ranging
Make a small but important change in the way that aux ranges are
calculated in topcat's PlotPanel.  Try where possible to use the
same surface for ranging as one which might have been used to do
the last plot on, rather than one that might be a bit different.
This makes it possible for plotters to use cached plans to avoid
duplicating work at the ranging stage.

At this commit, that's not used anywhere.