ttools: use cached plans for SkyDensityPlotter aux ranging
The SkyDensity plotter now attempts to use cached plan objects
generated during previous plots to get the new aux range.
This means that some recalculations can be avoided, so for
instance changing the colour map does not require a new scan
of the data.  The current ranging implementation still scales
with plot screen pixel count though, so it's not instantaneous
either.