ttools: improve strategy for HealpixPlotter tile painting
I experimented quite a bit with tweaking how tiles are rendered.
Painting is often faster, but for large tiles the shapes are wrong
near the edge of the sky (especially Aitoff).

I tried fixing SkySurfaceTiler so that it gives shapes which are
composed of lots of sub-tiles.  That works, but it doesn't seem
faster than the resampling, so don't bother doing it.

So if the tiles are quite big and no sky edges are visible, paint;
otherwise resample.  I needed to add more SkySurface methods
(previous commit) to support working out the size of the tiles.

This also (mostly) fixes a bug where the colour map was not getting
assigned properly for all-sky plots; if the sky edges were visible
the tile enumeration defaulted to all tiles, which meant that
the aux range used all pixels, rather than just the visible ones.
Now the renderer that does that is not used when the sky edges
are visible, so the problem doesn't arise.