ttools: fix taplint behaviour for missing /tables endpoint
Use custom code in taplint to read the VODataService document
at the TAP service's /tables endpoint, rather than using the
utility method TableSetSaxHandler.readTableSet().

This enables more careful reporting of errors and warnings,
and in particular allows to report the case of an HTTP 404
code correctly; this now reports a Warning (W_TBNF) rather
than an Error (W_FLIO).  TAP permits a 404 at /tables, so
the previous behaviour was a taplint bug.