vo: rework presentation order for TapSchema metadata acquisition
The TapSchemaTapMetaReader, which uses TapSchemaInterrogator,
now uses numeric ranking columns to define the return (display)
sequence for some metadata items.  Currently, this is just used for
columns (defined by the optional TAP 1.1 "column_index" column
in TAP_SCHEMA.columns).  It's a one-liner to use table_index and
schema_index if desired, but I think it's better not to, since
the other metadata acquisition policies currently fix those
orders to be alphabetic for item name.