vo: add queryStdOnly parameter to MetaQuerier
Streamline TapSchemaInterrogator a bit.
Now (unlike what it used to do) it generally queries all the
columns from each metadata table, because it's interested in
finding non-standard metadata for columns, tables etc.
That requires a bit of overhead; there may be a bit more
retrieved metadata and it needs to do an exploratory query
to find out what columns are available.

But for some of those queries (e.g. foreign keys) it doesn't
use that non-standard metadata.  So in those cases, just query
for the known, standard columns.