vo: exclude some known metadata items in extras
Some of the metadata items that may be acquired when interrogating
TAP_SCHEMA tables are not included in the standard list because
they may be absent (e.g. because they are introduced only in TAP 1.1).
But they don't want to be included in the user-visible extras list
because their information has already been taken account of.
So remove them explicitly from the list of extras.