vo: fall back on query ordering of TAP metadata items
Where the rank and alphabetic ordering of TAP metadata items
like ColumnMeta and TableMeta do not indicate a well-defined
sequence (alphabetic may be inapplicable since we have decided
not to order alphabetically, as for ColumnMeta, and rank ordering
may be inapplicable since e.g. column_index columns may not be
supplied by the service), fall back on the order of rows supplied
by the service when we make the TAP_SCHEMA query.
In RDBMS terms this ordering is not well-defined (can't be guaranteed),
but in practice it often gives you a sensible order.