vo: return typed column Extras from TableSetSaxHandler
The TableSetSaxHandler reads non-standard attribute values to
get non-standard column metadata (ColumnMeta Extras) information.
This comes as strings, but it's beneficial to return them as
typed values.  Add some code to make a best guess what type
various sets of column metadata information should be, and
convert it to that before it makes it out of TableSetSaxHandler.
Then things like JTable display and sorting works better.