fits: always close ArrayDataInput streams in FITS table read
Fix a long-standing bug in FitsTableBuilder.  It was failing to
close ArrayDataInput streams following a successful (though not
an unsuccessful) table read.

The relevant code (only close the stream if the table is null)
looked weird, but dated back to when loaded BintableStarTables
hung onto the stream itself to provide the data (problematic
for various reasons).  Since 3f61695e7c2 (STIL 3.0-9) the
stream is always discarded on table load, meaning this adjustment
is definitely overdue since then (before that there was still
a problem with failure to release resources, there just wasn't
an easy fix).

This bug was tracked down and reported by vforchi as starjava github
issue #6, noting that it could lead to exhaustion of file descriptors
if you were opening a lot of FITS tables.  Thanks vforchi.