fits: implement Closeable in {Bintable,ColFits}StarTable
BintableStarTable and ColfitsStarTable now have close methods
(as specified by java.io.Closeable).  These cause the FileChannel
held by the table to be closed, thus freeing associated file
descriptors.  This is not part of the StarTable contract, and
for most StarTable implementations it's not necessary, but it's
provided here so that if you happen to know you're using one of
these StarTable instances that hold on to file descriptors,
and you're in danger of running out of file descriptors,
you can call it to tidy up.

This relates to github issue #7.