ttools: fix plot2 aux auto-scaling bug
Some Plotter implementations were getting the wrong limits for
auto-ranged aux colour maps if the scaling was logarithmic and
there were negative values present.  The visual effect was that
a range much too small (near the top of the actual range) was used,
so the colour-coded plot generally looked all the same colour.

This was down to using Ranges incorrectly - a range primed with
a minimum and maximum value isn't good enough, because it also
needs the minimum positive value for use with log scaling.

Rewrite BinList.Result API and implementations to allow a fix
for this.  This commit fixes the bug.  It may have (minor?)
negative effects on some ranging performance.  The API changes
(an improvement) also make it possible to do some other optimisations
concerning preparation and use of BinLists.  No such optimisations
are attempted in this commit though.