ttools: use HealpixPlotter painting code in SkyDensityPlotter
Refactor and reuse the code for plotting Healpix pixels on to a
SkySurface from HealpixPlotter in SkyDensityPlotter.
This (now shared) code is smarther than what was previously done
in SkyDensityPlotter; it decides based on things like tile sizes
whether to draw the tiles by resampling or by painting.
Painting can be quite a bit faster for large tiles.

As well as the efficiency improvements, this refactoring makes
both classes somewhat tidier/more comprehensible.