ttools: fix JyStilts table handling bug
A problem with the implementation and use of the parameter system
meant that you could specify different tables in a JyStilts plotting
(and maybe other?) task invocation but the data from only one of
the tables would be used.  This came from some deep magic in
DataSpec/DataStore and ConsumerTask.IdentifierTable, and was related
to the identification of different tables specified in a stilts
execution Environment.

Fixed by causing the table parameter's String value to be table.toString()
rather than table.getName().  This has fixed the current problem,
as long as different tables don't have the same toString() return
value.  Probably they won't, but this is not guaranteed, so it's
possible this will come up again.  A proper fix would require
some quite careful changes to the parameter system
(see Parameter.setValueFrom{String,Object}).