topcat: add plot2 Aux Lock toggle button
There's a new toggle button on the plot window toolbars,
also visible in the Aux Axis control Range tab.

This controls whether the Aux range (and other auxiliary ranges)
is calculated dynamically from the plotted data at every frame,
or whether it's locked to a fixed value (whatever the value was
when the toggle was set).