ttools: add univariate statistics plotter
Calculates mean and standard deviation for a single-coordinate
dataset, like a cross between LinearFit and Histogram,
and then represents the result by plotting a Gaussian curve.

The user-visible name of this plotter is currently "Gausssian".
That is what it looks like, though the calculation is not really
Gaussian at all.  So the name is questionable (could be "stats"?
"mean-sigma"?, but I think it's probably as good a choice as any
(comprehensible, and not too likely to be wanted in future
for a different purpose).

Does not currently notice normalisation settings, maybe it should.