ttools: remove spurious padding from EPS graphics output
For some reason, the EPS GraphicExporter added a fractional padding
round the edge of generated EPS graphics files of 5% of the figure
width-or-height.  I must have thought that was a good idea at some
point, but it means the EPS output is not consistent with the other
graphics formats (e.g. PDF), and generally makes the whole thing
harder to control.

Remove the padding.  It is possible this change of behaviour might
cause trouble for users who have worked round it (or even benefitted
from it) in the past.  But I consider the previous behaviour a bug
because of the inconsistency between different output formats.