vo: inform UwsJobListPanel JList on job status update
A JList in UwsJobListPanel contains an entry for each running job
and displays its status.  However, the JList was not getting
messaged when job status was updated, so the status update was
not visible in the list unless you clicked the relevant entry or
something.  Fix it by messaging ListDataListeners in the right place.