vo: keep trying UWS status reads after some network errors
When watching (polling or blocking) a UWS job, previously most
IOExceptions caused the monitoring to bail out in tapquery
(and topcat? not sure).  That may not be appropriate, since
something like a "Connection refused" can occur if the server is
undergoing a restart, after which UWS jobs started earlier
may well persist and later complete.

So fix it so that certain network errors (corresponding basically
to network outages or remote server downtime) are tolerated,
and UwsJob.waitForFinish keeps waiting/retrying rather than
failing.  Probably this ought to cope with HTTP 503s as well,
but currently doesn't for reasons of implementor laziness.