ttools: add TracePlotter
This is a plotter that plots a line through a given quantile of
the points present in each pixel column (or row) of a 2d point cloud.
Added at the request of Uli Bastian for AGIS.

A few things missing.  Some kind of smoothing is probably desirable.

I initially called this plot Trace, but following Uli's advice
I have renamed it Quantile, but I'm still not sure.