ttools: stats plotter now senses logarithmic axis
If the X axis is logarithmic, Stats1Plotter now calculates and
reports mean and standard deviation for the data logarithms
rather than their values.  This means that the fit shows up
Gaussian on the logarithmic axis, rather than skewed - probably
what's required.  Since the meaning of this is becoming less
obvious, report the fitting equation as well as the coefficients.