ttools: add normalisation options to Stats1Plotter
There is now a config key for normalisation and bin size in
Stats1Plotter.  Despite the fact it doesn't have any bins.
Setting the bin size will cause it to scale the curve to follow
the line of a histogram which does have bins of a given size,
which makes sense if you're viewing it against a histogram.

The ReportMap is also changed somewhat; instead of an equation
for the curve, a function is emitted, which can be plugged
directly into the FunctionPlotter if required.