xdoc: use JIMI not JAI for FOP PNG processing
Recent changes move towards using PNGs rather than GIFs in the
topcat and stilts user documents.  PNGs are less outdated, and
offer less restricted colour ranges, variable transparency etc,
so this is worthwhile.

However, the FOP processing of XML documents (topcat/src/docs/sun253.xml,
and to a lesser extent ttools/src/docs/sun256.xml) runs into problems
when PNGs are used instead of GIFs.  Apache FOP is the XSLT-FO
processor used to convert the XML source (via xml-fo) into PDFs.

The version of FOP that's in use here (XDOC package) is 0.20.5.
That is very ancient, and dates from a time when PNG support was
not part of the JRE.  Hence it uses external, unbundled components
to handle PNGs, either JAI (Java Advanced Imaging) or JIMI
(some kind of ImageIO precusor??).
Unfortunately, the JAI implementation seems to be badly broken,
and attempting to use FOP+JAI to do this processing results in
many errors; complaints include stacktrace elements like this:

   ...
   Caused by: com.sun.media.jai.codecimpl.util.ImagingException: Error reading PNG header.
           ... 41 more
   Caused by: java.lang.RuntimeException: PNG magic number not found.
           at com.sun.media.jai.codecimpl.PNGImage.<init>(PNGImageDecoder.java:415)
       com.sun.media.jai.codecimpl.util.ImagingException: Error reading PNG header.
    ...

[Note: at time of writing, this PDF generation is not performed as
part of the normal build/install targets in topcat
(since it's very slow, takes 3-4 minutes), it has to be done by
explicit use of the done using the "docs-pdf" target in
the TOPCAT package.  Currently, it generally only gets invoked
in preparation for a topcat public release.
So a normal build/install cycle will not show up this problem.]

So, we can address this issue by adding the JIMI library.
The jar file added here (to xdoc/src/lib) is jimi.jar, downloaded from:

   https://java.net/projects/feedpod/sources/svn/content/trunk/totsp/lib/jimi/Jimi/jimi.jar?rev=188

This is identical to the zip file JimiProClasses.zip found within
jimi1_0.zip, so presumably corresponds to JIMI 1.0.

However, we're still not done.  If JAI and JIMI are both present,
FOP uses JAI by preference.  JAI very likely will be present during
the build, since the SOG package requires it, so it's likely to
be in the JDK's extensions directory.  We have to make sure it's
not picked up at FOP runtime, so in the topcat/ttools build files
we tweak the java.ext.dirs path to avoid seeing JAI at runtime, thus
leaving the way clear for JIMI.

The clean way to address this would be to upgrade to a more recent
version of FOP (at time of writing, it's at 2.1).  This handles
PNGs without requiring an external library, and it's also much
much faster.  But the behaviour is different - at a quick glance
image scaling needs changing, some justification looks wrong,
and it does that thing where the colour and antialising is all
wrong on pages that contain images.  Not clear how much effort
it would take to tackle these things.  But if gets done one day,
JIMI and the acrobatics in the topcat/ttools build files could
(and should) be retired.