ttools: add grid padding to GridPlotter
When calculating the grid data for plots, GridPlotter now calculates
over the edge of the visible area by some fraction, which is
generally not very expensive to do.
This means that the same plot plan can be used if the surface
has navigated to a different, but not too different position,
so it improves pan/zoom interactive performance significantly.