topcat: improve column assignment guesses in ConcatWindow
Improve the guessing in ConcatWindow about which columns in one
table match up with which other ones.

Previously, a uniquely matched UCD *and* a uniquely matched name
were required to match columns.  Now, in order we look for a
uniquely matched name, or UCD, or utype, any will do.
This logic is now in ColumnDataComboBoxModel rather than
ConcatWindow, since in principle it could be used elsewhere.