ttools: fix SkyDensity aux ranging bug
The level value used for identifying a suitable plan was the
level specification (could be negative to indicated a relative
value), but it should have been the level itself (always non-negative).
Fix it.  This allows a suitable plan to be used during aux ranging
when a relative level is in effect - performance improvement in
that case.