ttools: fix different SkyDensity aux ranging bug
Don't try to be so clever when calculating the aux range for
SkyDensity.  I was trying to bin only those values that are
visible on the current surface when doing range determination,
but not doing it carefully enough, so the bin result was not
properly constructed and so the aux range was wrong compared
to that used during the actual plot.  Now just use the same
procedure in both cases to build the bin result; it's missing
an opportunity for optimisation, but there's less to go wrong.
It gets the ranging about right now.