ttools: add ChooserColorSpecifier
Implement a new Specifier<Color> based on a JColorChooser.
This provides a much wider range of possible colours than
the existing ColorComboBox.

At present it is in the form of a small button that pops up
a dialogue when activated.

At this commit, the question of how to integrate this with the
existing ColorConfigKey is not addressed.